Moved route handlers to separate route files in `src/routes/`
